remove type duplicates; rearange types
