+++ packages/core/resolve-core/src/aggregate/aggregate.ts
@@ -1,7 +1,7 @@
 import { AggregateDomain } from './types'
 import { getAggregatesInteropBuilder } from './get-aggregates-interop-builder'
 import { validateAggregate } from './validate-aggregate'
-import { AggregateMeta } from '../types'
+import { AggregateMeta } from '../type-definitions/runtime'
 
 export const initAggregateDomain = (
   rawAggregates: AggregateMeta[]
