+++ packages/core/resolve-core/src/read-model/read-model.ts
@@ -1,7 +1,7 @@
 import { ReadModelDomain } from './types'
 import { getReadModelsInteropBuilder } from './get-read-models-interop-builder'
 import { validateReadModel } from './validate-read-model'
-import { ReadModelMeta } from '../types'
+import { ReadModelMeta } from '../type-definitions/runtime'
 
 export const initReadModelDomain = (
   rawReadModels: ReadModelMeta[]
