+++ packages/core/resolve-core/src/saga/build-scheduler-projection.ts
@@ -1,4 +1,4 @@
-import { SagaEventHandlers } from '../core-types'
+import { SagaEventHandlers } from '../type-definitions/core'
 import getLog from '../get-log'
 import {
   SchedulerEventTypes,
