+++ packages/core/resolve-core/src/saga/validate-saga.ts
@@ -1,4 +1,4 @@
-import { SagaMeta } from '../types'
+import { SagaMeta } from '../type-definitions/runtime'
 
 // FIXME: add validation (io-ts?)
 export const validateSaga = (saga: SagaMeta) => saga
