+++ packages/core/resolve-core/src/view-model/validate-view-model.ts
@@ -1,4 +1,4 @@
-import { ViewModelMeta } from '../types'
+import { ViewModelMeta } from '../type-definitions/runtime'
 
 // FIXME: add validation (io-ts?)
 export const validateViewModel = (viewModel: ViewModelMeta) => viewModel
