+++ packages/core/resolve-core/src/view-model/view-model.ts
@@ -1,7 +1,7 @@
 import { ViewModelDomain } from './types'
 import { getViewModelsInteropBuilder } from './get-view-models-interop-builder'
 import { validateViewModel } from './validate-view-model'
-import { ViewModelMeta } from '../types'
+import { ViewModelMeta } from '../type-definitions/runtime'
 
 export const initViewModelDomain = (
   rawViewModels: ViewModelMeta[]
