+++ packages/core/resolve-runtime/src/common/query/types.ts
@@ -1,5 +1,4 @@
 import {
-  IS_BUILT_IN,
   Monitoring,
   ReadModelInteropMap,
   SagaInteropMap,
@@ -36,34 +35,11 @@ export type WrapViewModelOptions = WrapModelOptions & {
   interop: ViewModelInterop
 }
 
-export type EventStoreAdapter = {
-  loadEvents: Function
-  getNextCursor: Function
-}
-
-export interface Serializer {
-  (state: any, jwt: string): string
-  [IS_BUILT_IN]?: boolean
-}
-
-export interface Deserializer {
-  (data: string): any
-  [IS_BUILT_IN]?: boolean
-}
-
 export type SerializedError = {
   name: string | null
   code: string | null
   message: string
   stack: string
-}
-
-export type ReadModelMeta = {
-  name: string
-  resolvers: { [key: string]: any }
-  projection: { [key: string]: Function }
-  connectorName: string
-  encryption: Function
 }
 
 export type ReadModelPool = {
@@ -76,16 +52,6 @@ export type ReadModelPool = {
   getVacantTimeInMillis: Function
   monitoring?: Monitoring
   provideLedger: (ledger: any) => Promise<void>
-}
-
-export type ViewModelMeta = {
-  name: string
-  invariantHash: string
-  deserializeState: Deserializer
-  serializeState: Serializer
-  projection: { [key: string]: Function }
-  resolver: Function
-  encryption: Function
 }
 
 export type ViewModelPool = {
